Remove template-utils as a dependency
The library hasn't been touched since before I created this project,
and the version number published on PyPI isn't considered a valid
format so tox was choking on it.

Remove the library, and remove the `markup_type` fields allowing the
user to select which markup language they are using.  Markdown is now
mandatory.